Rearranged the type parameters of Builder
The Hasher should go first because it is the only type parameter that
can't be inferred for Builder::new(). Then the leaf extractor, because
it's expected to be generic in the calling code more often than the
input type; when not generic, some callers might still prefer to specify
leaf::NoData for clarity. This also corresponds to the parameter order in
Builder::from_hasher_leaf_data().