Provide a separate ExtractData impl for plain fn
leaf::extract_with() now returns its plain fn parameter, so it
provides a type system gate for ensuring a lambda expression is
realised as a plain function, such that it can be used with
Builder::build_balanced_from().

For using real closures to extract leaf data (but not with
Builder::build_balanced_from()), a constructor
ExtractFn::with() is provided.