Refactored Hasher, added ByteDigestHasher
As node hashing does not depend on the input type, it is factored
out to a supertrait NodeHasher. This allows for less generic bloat
in the impls.

In mod digest, this allows a code-reusing implementation of a new
type, ByteDigestHasher, which accepts any kind of bytestream hash
functions directly, without the `Endian` layer that is not needed
for the byte slice convertible input this hash extractor supports.