Relax the trait bounds on construction methods to the minimum
Even if the minimal necessary bounds allow, in principle, creation
of unusable objects, intersecting bounds from all useful impls is
a maintenance burden, and the extra bounds get forced on the API
user at the construction site which may not be what they want.

Some of the type restrictions, instead, need to be imposed on the
whole types in order for them to make sense. Like, the
`DigestHasher` is only valid with a `NodeHasher` implementation type
that has the same output as the digest function.