A big redesign of Builders
No &mut self, internal data accumulation, or `self`-sacrificing methods
and the resulting cloning.
Instead, the Builder types have &self methods to build any number of
trees that are always managed externally. Great simplification ensues,
ExtractFn can be used for recursive tree builds in the sequential Builder,
and the APIs of both Builders use the same method names.

Make sure the built trees always take a well-defined form where
the depth of all leaves is the same.
This requires the Hasher will have to hash nodes with one child,
which it already did, and potentailly treat it as a special case.